@@ -1,13 +1,9 @@
 module github.com/unifi-poller/webserver
 
-go 1.14
+go 1.15
 
 require (
-	github.com/gorilla/mux v1.7.4
-	github.com/pkg/errors v0.9.1
-	github.com/prometheus/client_golang v1.7.1 // indirect
-	github.com/unifi-poller/poller v0.0.8-0.20200628131550-26430cac16c1
-	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
-	golang.org/x/sys v0.0.0-20200625212154-ddb9806d33ae // indirect
-	google.golang.org/protobuf v1.25.0 // indirect
+	github.com/gorilla/mux v1.8.0
+	github.com/unifi-poller/poller v0.0.8
+	golang.org/x/crypto v0.0.0-20210220033148-5ea612d1eb83
 )
